+++ src/tld/__init__.py
@@ -1,4 +1,5 @@
 from .utils import (
+    get_fld,
     get_tld,
     get_tld_names,
     Result,
@@ -6,11 +7,12 @@ from .utils import (
 )
 
 __title__ = 'tld'
-__version__ = '0.7.10'
+__version__ = '0.8'
 __author__ = 'Artur Barseghyan'
 __copyright__ = '2013-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
+    'get_fld',
     'get_tld',
     'get_tld_names',
     'Result',