+++ src/tld/tests.py
@@ -7,8 +7,14 @@ import unittest
 
 from . import defaults
 from .conf import get_setting, set_setting
+from .exceptions import (
+    TldBadUrl,
+    TldDomainNotFound,
+    TldImproperlyConfigured,
+    TldIOError,
+)
 from .helpers import project_dir
-from .utils import get_tld, update_tld_names
+from .utils import get_fld, get_tld, update_tld_names
 
 __title__ = 'tld.tests'
 __author__ = 'Artur Barseghyan'
@@ -60,19 +66,21 @@ class TldTest(unittest.TestCase):
         self.good_patterns = [
             {
                 'url': 'http://www.google.co.uk',
-                'tld': 'google.co.uk',
+                'fld': 'google.co.uk',
                 'subdomain': 'www',
                 'domain': 'google',
                 'suffix': 'co.uk',
-                'kwargs': {'fail_silently': True}
+                'tld': 'co.uk',
+                'kwargs': {'fail_silently': True},
             },
             {
                 'url': 'http://www.v2.google.co.uk',
-                'tld': 'google.co.uk',
+                'fld': 'google.co.uk',
                 'subdomain': 'www.v2',
                 'domain': 'google',
                 'suffix': 'co.uk',
-                'kwargs': {'fail_silently': True}
+                'tld': 'co.uk',
+                'kwargs': {'fail_silently': True},
             },
             # No longer valid
             # {
@@ -84,73 +92,174 @@ class TldTest(unittest.TestCase):
             # },
             {
                 'url': u'http://хром.гугл.рф',
-                'tld': u'гугл.рф',
+                'fld': u'гугл.рф',
                 'subdomain': u'хром',
                 'domain': u'гугл',
                 'suffix': u'рф',
-                'kwargs': {'fail_silently': True}
+                'tld': u'рф',
+                'kwargs': {'fail_silently': True},
             },
             {
                 'url': 'http://www.google.co.uk:8001/lorem-ipsum/',
-                'tld': 'google.co.uk',
+                'fld': 'google.co.uk',
                 'subdomain': 'www',
                 'domain': 'google',
                 'suffix': 'co.uk',
-                'kwargs': {'fail_silently': True}
+                'tld': 'co.uk',
+                'kwargs': {'fail_silently': True},
             },
             {
                 'url': 'http://www.me.cloudfront.net',
-                'tld': 'me.cloudfront.net',
+                'fld': 'me.cloudfront.net',
                 'subdomain': 'www',
                 'domain': 'me',
                 'suffix': 'cloudfront.net',
-                'kwargs': {'fail_silently': True}
+                'tld': 'cloudfront.net',
+                'kwargs': {'fail_silently': True},
             },
             {
                 'url': 'http://www.v2.forum.tech.google.co.uk:8001/'
                        'lorem-ipsum/',
-                'tld': 'google.co.uk',
+                'fld': 'google.co.uk',
                 'subdomain': 'www.v2.forum.tech',
                 'domain': 'google',
                 'suffix': 'co.uk',
-                'kwargs': {'fail_silently': True}
+                'tld': 'co.uk',
+                'kwargs': {'fail_silently': True},
             },
             {
                 'url': 'https://pantheon.io/',
-                'tld': 'pantheon.io',
+                'fld': 'pantheon.io',
                 'subdomain': '',
                 'domain': 'pantheon',
                 'suffix': 'io',
-                'kwargs': {'fail_silently': True}
+                'tld': 'io',
+                'kwargs': {'fail_silently': True},
             },
             {
                 'url': 'v2.www.google.com',
-                'tld': 'google.com',
+                'fld': 'google.com',
                 'subdomain': 'v2.www',
                 'domain': 'google',
                 'suffix': 'com',
-                'kwargs': {'fail_silently': True, 'fix_protocol': True}
+                'tld': 'com',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
             },
             {
                 'url': '//v2.www.google.com',
-                'tld': 'google.com',
+                'fld': 'google.com',
                 'subdomain': 'v2.www',
                 'domain': 'google',
                 'suffix': 'com',
-                'kwargs': {'fail_silently': True, 'fix_protocol': True}
-            }
+                'tld': 'com',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
+            },
+            {
+                'url': 'http://foo@bar.com',
+                'fld': 'bar.com',
+                'subdomain': '',
+                'domain': 'bar',
+                'suffix': 'com',
+                'tld': 'com',
+                'kwargs': {'fail_silently': True},
+            },
+            {
+                'url': 'http://user:foo@bar.com',
+                'fld': 'bar.com',
+                'subdomain': '',
+                'domain': 'bar',
+                'suffix': 'com',
+                'tld': 'com',
+                'kwargs': {'fail_silently': True},
+            },
+            {
+                'url': 'https://faguoren.xn--fiqs8s',
+                'fld': 'faguoren.xn--fiqs8s',
+                'subdomain': '',
+                'domain': 'faguoren',
+                'suffix': 'xn--fiqs8s',
+                'tld': 'xn--fiqs8s',
+                'kwargs': {'fail_silently': True},
+            },
+            {
+                'url': 'blogs.lemonde.paris',
+                'fld': 'lemonde.paris',
+                'subdomain': 'blogs',
+                'domain': 'lemonde',
+                'suffix': 'paris',
+                'tld': 'paris',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
+            },
+            {
+                'url': 'axel.brighton.ac.uk',
+                'fld': 'brighton.ac.uk',
+                'subdomain': 'axel',
+                'domain': 'brighton',
+                'suffix': 'ac.uk',
+                'tld': 'ac.uk',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
+            },
+            {
+                'url': 'm.fr.blogspot.com.au',
+                'fld': 'fr.blogspot.com.au',
+                'subdomain': 'm',
+                'domain': 'fr',
+                'suffix': 'blogspot.com.au',
+                'tld': 'blogspot.com.au',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
+            },
+            {
+                'url': u'help.www.福岡.jp',
+                'fld': u'www.福岡.jp',
+                'subdomain': 'help',
+                'domain': 'www',
+                'suffix': u'福岡.jp',
+                'tld': u'福岡.jp',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
+            },
+            {
+                'url': u'syria.arabic.variant.سوريا',
+                'fld': u'variant.سوريا',
+                'subdomain': 'syria.arabic',
+                'domain': 'variant',
+                'suffix': u'سوريا',
+                'tld': u'سوريا',
+                'kwargs': {'fail_silently': True, 'fix_protocol': True},
+            },
         ]
 
-        self.bad_patterns = [
-            'v2.www.google.com',
-            '/index.php?a=1&b=2',
-            'http://www.tld.doesnotexist'
-        ]
+        self.bad_patterns = {
+            'v2.www.google.com': {
+                'exception': TldBadUrl,
+            },
+            '/index.php?a=1&b=2': {
+                'exception': TldBadUrl,
+            },
+            'http://www.tld.doesnotexist': {
+                'exception': TldDomainNotFound,
+            },
+            'https://2001:0db8:0000:85a3:0000:0000:ac1f:8001': {
+                'exception': TldDomainNotFound,
+            },
+            'http://192.169.1.1': {
+                'exception': TldDomainNotFound,
+            },
+            'http://localhost:8080': {
+                'exception': TldDomainNotFound,
+            },
+            'https://localhost': {
+                'exception': TldDomainNotFound,
+            },
+            'https://localhost2': {
+                'exception': TldImproperlyConfigured,
+                'kwargs': {'search_public': False, 'search_private': False,}
+            },
+        }
 
     @log_info
     def test_0_tld_names_loaded(self):
         """Test if tld names are loaded."""
-        get_tld('http://www.google.co.uk')
+        get_fld('http://www.google.co.uk')
         from .utils import tld_names
         res = len(tld_names) > 0
         self.assertTrue(res)
@@ -164,21 +273,21 @@ class TldTest(unittest.TestCase):
         return res
 
     @log_info
-    def test_2_good_patterns_pass(self):
+    def test_2_fld_good_patterns_pass(self):
         """Test good URL patterns."""
         res = []
         for data in self.good_patterns:
-            _res = get_tld(data['url'], **data['kwargs'])
-            self.assertEqual(_res, data['tld'])
+            _res = get_fld(data['url'], **data['kwargs'])
+            self.assertEqual(_res, data['fld'])
             res.append(_res)
         return res
 
     @log_info
-    def test_3_bad_patterns_pass(self):
+    def test_3_fld_bad_patterns_pass(self):
         """Test bad URL patterns."""
         res = []
-        for url in self.bad_patterns:
-            _res = get_tld(url, fail_silently=True)
+        for url, params in self.bad_patterns.items():
+            _res = get_fld(url, fail_silently=True)
             self.assertEqual(_res, None)
             res.append(_res)
         return res
@@ -199,7 +308,7 @@ class TldTest(unittest.TestCase):
         return override_settings()
 
     @log_info
-    def test_5_good_patterns_pass_parsed_object(self):
+    def test_5_tld_good_patterns_pass_parsed_object(self):
         """Test good URL patterns."""
         res = []
         for data in self.good_patterns:
@@ -210,6 +319,7 @@ class TldTest(unittest.TestCase):
             self.assertEqual(_res.subdomain, data['subdomain'])
             self.assertEqual(_res.domain, data['domain'])
             self.assertEqual(_res.suffix, data['suffix'])
+            self.assertEqual(_res.fld, data['fld'])
             res.append(_res)
         return res
 
@@ -221,6 +331,73 @@ class TldTest(unittest.TestCase):
         modified = project_dir('dummy.txt')
         self.assertNotEqual(default, modified)
         self.assertEqual(modified, os.path.abspath('/tmp/test/dummy.txt'))
+
+    @log_info
+    def test_7_public_private(self):
+        res = get_fld(
+            'http://silly.cc.ua',
+            fail_silently=True,
+            search_private=False
+        )
+
+        self.assertEqual(res, None)
+
+        res = get_fld(
+            'http://silly.cc.ua',
+            fail_silently=True,
+            search_private=True
+        )
+
+        self.assertEqual(res, 'silly.cc.ua')
+
+        res = get_fld(
+            'mercy.compute.amazonaws.com',
+            fail_silently=True,
+            search_private=False,
+            fix_protocol=True
+        )
+
+        self.assertEqual(res, None)
+
+        res = get_fld(
+            'http://whatever.com',
+            fail_silently=True,
+            search_public=False
+        )
+
+        self.assertEqual(res, None)
+
+    @log_info
+    def test_8_fld_bad_patterns_exceptions(self):
+        """Test exceptions."""
+        res = []
+        for url, params in self.bad_patterns.items():
+            kwargs = params['kwargs'] if 'kwargs' in params else {}
+            kwargs.update({'fail_silently': False})
+            with self.assertRaises(params['exception']):
+                _res = get_fld(url, **kwargs)
+                res.append(_res)
+        return res
+
+    @log_info
+    def test_9_tld_good_patterns_pass(self):
+        """Test `get_tld` good URL patterns."""
+        res = []
+        for data in self.good_patterns:
+            _res = get_tld(data['url'], **data['kwargs'])
+            self.assertEqual(_res, data['tld'])
+            res.append(_res)
+        return res
+
+    @log_info
+    def test_10_tld_bad_patterns_pass(self):
+        """Test `get_tld` bad URL patterns."""
+        res = []
+        for url, params in self.bad_patterns.items():
+            _res = get_tld(url, fail_silently=True)
+            self.assertEqual(_res, None)
+            res.append(_res)
+        return res
 
 
 if __name__ == '__main__':