+++ src/tld/utils.py
@@ -3,11 +3,16 @@ from __future__ import unicode_literals
 import codecs
 
 from six import PY3, text_type
-from six.moves.urllib.parse import urlparse
+from six.moves.urllib.parse import urlsplit
 from six.moves.urllib.request import urlopen
 
 from .conf import get_setting
-from .exceptions import TldIOError, TldDomainNotFound, TldBadUrl
+from .exceptions import (
+    TldBadUrl,
+    TldDomainNotFound,
+    TldImproperlyConfigured,
+    TldIOError,
+)
 from .helpers import project_dir
 
 __title__ = 'tld.utils'
@@ -15,49 +20,104 @@ __author__ = 'Artur Barseghyan'
 __copyright__ = '2013-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
+    'get_fld',
     'get_tld',
     'get_tld_names',
+    'process_url',
     'Result',
     'update_tld_names',
 )
 
-tld_names = set()
+tld_names = None
 
 
 class Result(object):
     """Container."""
 
-    __slots__ = ('subdomain', 'domain', 'suffix', '__tld')
+    __slots__ = ('subdomain', 'domain', 'tld', '__fld')
 
-    def __init__(self, subdomain, domain, suffix):
+    def __init__(self, tld, domain, subdomain):
         self.subdomain = subdomain
         self.domain = domain
-        self.suffix = suffix
-        self.__tld = "{0}.{1}".format(self.domain, self.suffix)
-
-    @property
-    def tld(self):
-        """TLD."""
-        return self.__tld
+        self.tld = tld
+        self.__fld = "{0}.{1}".format(self.domain, self.tld)
 
     @property
     def extension(self):
-        """Alias of ``suffix``.
+        """Alias of ``tld``.
 
         :return str:
         """
-        return self.suffix
+        return self.tld
+    suffix = extension
+
+    @property
+    def fld(self):
+        """First level domain.
+
+        :return:
+        """
+        return self.__fld
 
     def __unicode__(self):
         if PY3:
-            return self.__tld
+            return self.tld
         else:
             try:
-                return self.__tld.encode('utf8')
+                return self.tld.encode('utf8')
             except UnicodeEncodeError:
-                return self.__tld
+                return self.tld
     __repr__ = __unicode__
     __str__ = __unicode__
+
+
+class TrieNode(object):
+    """Class representing a single Trie node."""
+
+    __slots__ = ('children', 'leaf', 'private')
+
+    def __init__(self):
+        self.children = None
+        self.leaf = False
+        self.private = False
+
+
+class Trie(object):
+    """An adhoc Trie data structure to store tlds in reverse notation order."""
+
+    def __init__(self):
+        self.root = TrieNode()
+        self.__nodes = 0
+
+    def __len__(self):
+        return self.__nodes
+
+    def add(self, tld, private=False):
+        node = self.root
+
+        # Iterating over the tld parts in reverse order
+        for part in reversed(tld.split('.')):
+
+            # To save up some RAM, we initialize the children dict only
+            # when strictly necessary
+            if node.children is None:
+                node.children = {}
+
+            child = node.children.get(part)
+
+            if child is None:
+                child = TrieNode()
+
+            node.children[part] = child
+
+            node = child
+
+        node.leaf = True
+
+        if private:
+            node.private = True
+
+        self.__nodes += 1
 
 
 def update_tld_names(fail_silently=False):
@@ -112,7 +172,7 @@ def get_tld_names(fail_silently=False, retry_count=0):
     global tld_names
 
     # If already loaded, return
-    if len(tld_names):
+    if tld_names is not None:
         return tld_names
 
     local_file = None
@@ -121,10 +181,23 @@ def get_tld_names(fail_silently=False, retry_count=0):
         local_file = codecs.open(project_dir(tld_names_local_path),
                                  'r',
                                  encoding='utf8')
+        tld_names = Trie()
         # Make a list of it all, strip all garbage
-        tld_names = set([u'{0}'.format(line.strip())
-                         for line
-                         in local_file if line[0] not in '/\n'])
+        private_section = False
+
+        for line in local_file:
+            if '===BEGIN PRIVATE DOMAINS===' in line:
+                private_section = True
+
+            # Puny code tlds
+            if '// xn--' in line:
+                line = line.split(' (', 1)[0][3:]
+
+            if line[0] == '/' or line[0] == '\n':
+                continue
+
+            tld_names.add(u'{0}'.format(line.strip()), private=private_section)
+
         local_file.close()
     except IOError as err:
         update_tld_names()  # Grab the file
@@ -145,11 +218,162 @@ def get_tld_names(fail_silently=False, retry_count=0):
     return tld_names
 
 
+def process_url(url,
+                active_only=False,
+                fail_silently=False,
+                fix_protocol=False,
+                search_public=True,
+                search_private=True):
+    """Process URL.
+
+    :param url:
+    :param active_only:
+    :param fail_silently:
+    :param fix_protocol:
+    :param search_public:
+    :param search_private:
+    :return:
+    """
+    if not (search_public or search_private):
+        raise TldImproperlyConfigured(
+            "Either `search_public` or `search_private` (or both) shall be "
+            "set to True."
+        )
+
+    url = url.lower()
+
+    if fix_protocol:
+        if (
+            not url.startswith('//')
+            and not (url.startswith('http://') or url.startswith('https://'))
+        ):
+            url = 'https://{}'.format(url)
+
+    tld_names = get_tld_names(fail_silently=fail_silently)  # Init
+
+    # Get (sub) domain name
+    domain_name = urlsplit(url).netloc
+
+    # Handling auth
+    if '@' in domain_name:
+        domain_name = domain_name.split('@', 1)[-1]
+
+    # Handling port
+    domain_name = domain_name.split(':', 1)[0]
+
+    if not domain_name:
+        if fail_silently:
+            return None, None
+        else:
+            raise TldBadUrl(url=url)
+
+    domain_parts = domain_name.split('.')
+
+    # Now we query our Trie iterating on the domain parts in reverse order
+    node = tld_names.root
+    tld_length = 0
+    for i in reversed(range(len(domain_parts))):
+        part = domain_parts[i]
+
+        # Cannot go deeper
+        if node.children is None:
+            break
+
+        child = node.children.get(part)
+
+        # Wildcards
+        if child is None:
+            child = node.children.get('*')
+
+        # Inactive
+        if active_only is False and child is None:
+            child = node.children.get('!{0}'.format(part))
+
+        # If the current part is not in current node's children, we can stop
+        if child is None:
+            break
+
+        # Else we move deeper and increment our tld offset
+        tld_length += 1
+        node = child
+
+    # Checking the node we finished on is a leaf and is one we allow
+    if (
+        (not node.leaf) or
+        (not search_public and not node.private) or
+        (not search_private and node.private)
+    ):
+        if fail_silently:
+            return None, None
+        else:
+            raise TldDomainNotFound(domain_name=domain_name)
+
+    non_zero_i = max(1, len(domain_parts) - tld_length)
+
+    return domain_parts, non_zero_i
+
+
+def get_fld(url,
+            active_only=False,
+            fail_silently=False,
+            fix_protocol=False,
+            search_public=True,
+            search_private=True,
+            **kwargs):
+    """Extract the first level domain.
+
+    Extract the top level domain based on the mozilla's effective TLD names
+    dat file. Returns a string. May throw ``TldBadUrl`` or
+    ``TldDomainNotFound`` exceptions if there's bad URL provided or no TLD
+    match found respectively.
+
+    :param url: URL to get top level domain from.
+    :param active_only: If set to True, only active patterns are matched.
+    :param fail_silently: If set to True, no exceptions are raised and None
+        is returned on failure.
+    :param fix_protocol: If set to True, missing or wrong protocol is
+        ignored (https is appended instead).
+    :param search_public: If set to True, search in public domains.
+    :param search_private: If set to True, search in private domains.
+    :type url: str
+    :type active_only: bool
+    :type fail_silently: bool
+    :type fix_protocol: bool
+    :type search_public: bool
+    :type search_private: bool
+    :return: String with top level domain (if ``as_object`` argument
+        is set to False) or a ``tld.utils.Result`` object (if ``as_object``
+        argument is set to True); returns None on failure.
+    :rtype: str
+    """
+    if 'as_object' in kwargs:
+        raise TldImproperlyConfigured(
+            "`as_object` argument is deprecated for `get_fld`. Use `get_tld` "
+            "instead."
+        )
+
+    domain_parts, non_zero_i = process_url(
+        url=url,
+        active_only=active_only,
+        fail_silently=fail_silently,
+        fix_protocol=fix_protocol,
+        search_public=search_public,
+        search_private=search_private
+    )
+
+    if domain_parts is None:
+        return None
+
+    return text_type(".").join(domain_parts[non_zero_i-1:])
+
+
 def get_tld(url,
             active_only=False,
             fail_silently=False,
             as_object=False,
-            fix_protocol=False):
+            fix_protocol=False,
+            search_public=True,
+            search_private=True):
     """Extract the top level domain.
 
     Extract the top level domain based on the mozilla's effective TLD names
@@ -165,76 +389,43 @@ def get_tld(url,
         ``domain``, ``suffix`` and ``tld`` properties.
     :param fix_protocol: If set to True, missing or wrong protocol is
         ignored (https is appended instead).
+    :param search_public: If set to True, search in public domains.
+    :param search_private: If set to True, search in private domains.
     :type url: str
     :type active_only: bool
     :type fail_silently: bool
     :type as_object: bool
     :type fix_protocol: bool
+    :type search_public: bool
+    :type search_private: bool
     :return: String with top level domain (if ``as_object`` argument
         is set to False) or a ``tld.utils.Result`` object (if ``as_object``
         argument is set to True); returns None on failure.
     :rtype: str
     """
-    url = url.lower()
+    domain_parts, non_zero_i = process_url(
+        url=url,
+        active_only=active_only,
+        fail_silently=fail_silently,
+        fix_protocol=fix_protocol,
+        search_public=search_public,
+        search_private=search_private
+    )
 
-    if fix_protocol:
-        if (
-            not url.startswith('//')
-            and not (url.startswith('http://') or url.startswith('https://'))
-        ):
-            url = 'https://{}'.format(url)
-
-    tld_names = get_tld_names(fail_silently=fail_silently)  # Init
-
-    # Get (sub) domain name
-    domain_name = urlparse(url).netloc.split(":", 1)[0]
-
-    if not domain_name:
-        if fail_silently:
-            return None
-        else:
-            raise TldBadUrl(url=url)
-
-    domain_parts = domain_name.split('.')
-
-    # Looping from much to less (for example if we have a domain named
-    # "v3.api.google.co.uk" we'll try "v3.api.google.co.uk", then
-    # "api.google.co.uk", then "api.google.co.uk", then "google.co.uk", then
-    # "co.uk" and finally "uk". If the last one does not match any TLDs, we
-    # throw a <TldDomainNotFound> exception.
-    for i in range(0, len(domain_parts)):
-        sliced_domain_parts = domain_parts[i:]
-
-        match = text_type('.').join(sliced_domain_parts)
-        wildcard_match = text_type('.').join(['*'] + sliced_domain_parts[1:])
-        inactive_match = text_type("!{0}").format(match)
-
-        # if not PY3:
-        #    try:
-        #        match = match.encode('utf8')
-        #        wildcard_match = wildcard_match.encode('utf8')
-        #        inactive_match = inactive_match.encode('utf8')
-        #    except UnicodeDecodeError as e:
-        #        pass
-
-        # Match tlds
-        if (match in tld_names or
-                wildcard_match in tld_names or
-                (active_only is False and inactive_match in tld_names)):
-            # if url contains only the TLD (without sub-domains) then entire
-            # domain should be returned.
-            non_zero_i = max(1, i)
-            if not as_object:
-                return text_type(".").join(domain_parts[non_zero_i-1:])
-            else:
-                subdomain = text_type(".").join(domain_parts[:non_zero_i-1])
-                domain = text_type(".").join(
-                    domain_parts[non_zero_i-1:non_zero_i]
-                )
-                suffix = text_type(".").join(domain_parts[non_zero_i:])
-                return Result(subdomain, domain, suffix)
-
-    if fail_silently:
+    if domain_parts is None:
         return None
-    else:
-        raise TldDomainNotFound(domain_name=domain_name)
+
+    if not as_object:
+        return text_type(".").join(domain_parts[non_zero_i:])
+
+    subdomain = text_type(".").join(domain_parts[:non_zero_i-1])
+    domain = text_type(".").join(
+        domain_parts[non_zero_i-1:non_zero_i]
+    )
+    _tld = text_type(".").join(domain_parts[non_zero_i:])
+
+    return Result(
+        subdomain=subdomain,
+        domain=domain,
+        tld=_tld
+    )